Append and prepend methods created
Create append and prepend on LinkedList class. Every node has a pointer,
to append a node to the linked list we go to each of the nodes and its
pointers (starting from LinkedList.head) until we find the node with
pointer == nil (the last one).

Separate main.rb into LinkedList.rb and Node.rb